+++ src/app/api/chat/[provider]/agentRuntime.ts
@@ -21,6 +21,7 @@ import {
   LobeOpenRouterAI,
   LobePerplexityAI,
   LobeRuntimeAI,
+  LobeTogetherAI,
   LobeZeroOneAI,
   LobeZhipuAI,
   ModelProvider,
@@ -181,6 +182,11 @@ class AgentRuntime {
         break;
       }
 
+      case ModelProvider.TogetherAI: {
+        runtimeModel = this.initTogetherAI(payload);
+        break;
+      }
+
       case ModelProvider.ZeroOne: {
         runtimeModel = this.initZeroOne(payload);
         break;
@@ -301,13 +307,19 @@ class AgentRuntime {
     return new LobeOpenRouterAI({ apiKey });
   }
 
+  private static initTogetherAI(payload: JWTPayload) {
+    const { TOGETHERAI_API_KEY } = getServerConfig();
+    const apiKey = apiKeyManager.pick(payload?.apiKey || TOGETHERAI_API_KEY);
+
+    return new LobeTogetherAI({ apiKey });
+  }
+
   private static initZeroOne(payload: JWTPayload) {
     const { ZEROONE_API_KEY } = getServerConfig();
     const apiKey = apiKeyManager.pick(payload?.apiKey || ZEROONE_API_KEY);
 
     return new LobeZeroOneAI({ apiKey });
   }
-
 }
 
 export default AgentRuntime;