+++ src/store/global/slices/settings/selectors/modelProvider.ts
@@ -12,6 +12,7 @@ import {
   OpenAIProvider,
   OpenRouterProvider,
   PerplexityProvider,
+  TogetherAIProvider,
   ZeroOneProvider,
   ZhiPuProvider,
 } from '@/config/modelProviders';
@@ -66,6 +67,9 @@ const groqAPIKey = (s: GlobalStore) => modelProvider(s).groq.apiKey;
 
 const enableOpenrouter = (s: GlobalStore) => modelProvider(s).openrouter.enabled;
 const openrouterAPIKey = (s: GlobalStore) => modelProvider(s).openrouter.apiKey;
+
+const enableTogetherAI = (s: GlobalStore) => modelProvider(s).togetherai.enabled;
+const togetheraiAPIKey = (s: GlobalStore) => modelProvider(s).togetherai.apiKey;
 
 const enableZeroone = (s: GlobalStore) => modelProvider(s).zeroone.enabled;
 const zerooneAPIKey = (s: GlobalStore) => modelProvider(s).zeroone.apiKey;
@@ -151,14 +155,25 @@ const modelSelectList = (s: GlobalStore): ModelProviderCard[] => {
 
   const openrouterModelString = [
     s.serverConfig.languageModel?.openrouter?.customModelName,
-    currentSettings(s).languageModel.openrouter.customModelName
+    currentSettings(s).languageModel.openrouter.customModelName,
   ]
     .filter(Boolean)
     .join(',');
-  
+
   const openrouterModelConfig = parseModelString(openrouterModelString);
-  
-  const openrouterChatModels = processChatModels(openrouterModelConfig, OpenRouterProvider.chatModels);
+
+  const openrouterChatModels = processChatModels(
+    openrouterModelConfig,
+    OpenRouterProvider.chatModels,
+  );
+
+  const togetheraiModelConfig = parseModelString(
+    currentSettings(s).languageModel.togetherai.customModelName,
+  );
+  const togetheraiChatModels = processChatModels(
+    togetheraiModelConfig,
+    TogetherAIProvider.chatModels,
+  );
 
   return [
     {
@@ -175,8 +190,9 @@ const modelSelectList = (s: GlobalStore): ModelProviderCard[] => {
     { ...GroqProvider, enabled: enableGroq(s) },
     { ...ZhiPuProvider, enabled: enableZhipu(s) },
     { ...MoonshotProvider, enabled: enableMoonshot(s) },
-    { ...OpenRouterProvider, chatModels: openrouterChatModels, enabled: enableOpenrouter(s)},
-    { ...ZeroOneProvider, enabled: enableZeroone(s) }
+    { ...OpenRouterProvider, chatModels: openrouterChatModels, enabled: enableOpenrouter(s) },
+    { ...ZeroOneProvider, enabled: enableZeroone(s) },
+    { ...TogetherAIProvider, chatModels: togetheraiChatModels, enabled: enableTogetherAI(s) },
   ];
 };
 
@@ -271,4 +287,8 @@ export const modelProviderSelectors = {
   // ZeroOne 零一万物
   enableZeroone,
   zerooneAPIKey,
+
+  // TogetherAI
+  enableTogetherAI,
+  togetheraiAPIKey,
 };